SP, JK2: g_savegame: silence compiler warning
.../code/game/g_savegame.cpp: In function ‘void
EnumerateFields(const save_field_t*, const byte*, unsigned int, int)’:
.../code/game/g_savegame.cpp:512:22: warning: comparison between
signed and unsigned integer expressions [-Wsign-compare]
    assert(pField->iOffset < iLen);
                      ^

iLen is a non-negative amount of memory, so size_t is a more
appropriate type than int.

+++ code/game/g_savegame.cpp
@@ -499,7 +499,7 @@ static void EnumerateField(const save_field_t *pField, const byte *pbBase)
 	}
 }
 
-static void EnumerateFields(const save_field_t *pFields, const byte *pbData, unsigned int ulChid, int iLen)
+static void EnumerateFields(const save_field_t *pFields, const byte *pbData, unsigned int ulChid, size_t iLen)
 {
 	strList = new std::list<sstring_t>;
